BUG: Power limiting did not work. Hub4Control need the inverter status code to relate to Fronius. Added its own settings to relay operating state information, mode and condition to dbus.
Added extra layer to only push change to dbus on change of value.
To be able to login into the inverter with a grid code is an important
step to be able to control the output power of the inverter. The inverter
DBUS node has an item showing the driver was able to login. State codes
will also be set accordingly to tell Hub4Control if the inverter can do
power control, i.e. login was successfull.